feat: find product and insert orderProduct using promise

+++ src/migration/1783230030078-create-table-order-product.ts
@@ -8,7 +8,7 @@ export class CreateTableOrderProduct1783230030078 implements MigrationInterface 
             order_id int NOT NULL,
             product_id int NOT NULL,
             amount int NOT NULL,
-            price int NOT NULL,
+            price numeric(10,2) NOT NULL,
             created_at timestamp without time zone DEFAULT now() NOT NULL,
             updated_at timestamp without time zone DEFAULT now() NOT NULL,
             primary key (id),
